Add a post step for removing packages
